I have completed the work to refactor the smiley translator so it can be used by the new forum port and the existing code. To do so I have moved it and some dependencies to utils and objects classes, added unit/acceptance tests and relocated existing tests.
I have modified the architecture in two ways. Firstly it worked on a first through the door loaded the list - this is now moved a application startup as it was taken some time to initialise. This may have been due to fact that it built a massive regular expression (including all the possible smileys) which tooks seconds to build. I have removed the regular expressions in favour of basic find and replace - it works and is performant in profanites so it is an easy fit here.