Fix for trailing ) in urls
The issue is that the c++ code looks for .,? and ) trailing and thinks its punctuation. I have left . and ? in place but removed ')'. 

This does mean that links with double )'s will have broken links but its very difficult to weed out.
e.g  this is a test with in brackets (at the end http://en.wikipedia.org/wiki/Sink_or_Swim_(song)) will be transform to this is a test with in brackets (at the end <LINK HREF=\"http://en.wikipedia.org/wiki/Sink_or_Swim_(song))\">http://en.wikipedia.org/wiki/Sink_or_Swim_(song))</LINK>
which isn't fully correct but...

In c# I have changed the regular expression to achieve the same - and added relevant unit tests