Putting back code Martin did to fix bug 312.  This email was sent on 15/12/2009: --------------------------------- Bug 312 - Handing bad urls
updateexmoderationevents.sql
Changed the logic - items were kept in the queue until the time since last attempt was > 24 hours. This meant items were kept in the event queue for a number of days whilst the timeout interval grew to > 24 hours. 
Items are now removed from the queue if the item has been in the queue for > 24 hours.
To test this functionality manually:
 
1. Put a bad url into the queue.
2. Check the url with the modferation history page.
3. The item should be marked as removed after 1 day.