I've finished reviewing forumgetthreadlist.  This review was triggered by changes in the last release that I was meant to review before going live.  I guess better late than never.
Anyway, that SP has been bugging me for a while as it seemed to be taking up more resources than it ought to, and that turned out to be the case.  It's also one of our most called SPs, so it's a good one to tune.

Firstly there was a missing index that I added that makes the code for the new cache freshness checks go quicker.  So instead of a partial index scan, it can do an index seek for the max LastUpdated value in the thread.

Secondly, a minor rewrite of one of the joins massively reduced the amount of CPU required:

I changed:
INNER JOIN ThreadEntries te1 WITH(NOLOCK) ON te1.ThreadID = t.ThreadID AND te1.PostIndex IN (SELECT MAX(PostIndex) FROM ThreadEntries tex WITH(NOLOCK) WHERE tex.ThreadID = t.ThreadID) 

To:
INNER JOIN ThreadEntries te1 WITH(NOLOCK) ON te1.ThreadID = t.ThreadID AND te1.PostIndex = (SELECT top 1 PostIndex FROM ThreadEntries tex WITH(NOLOCK) WHERE tex.ThreadID = t.ThreadID ORDER BY PostIndex DESC) 

The optimiser was getting it's estimations wildly wrong in the "MAX" version, and came up with a very expensive plan with a nested loop for every thread entry for every thread in the result set.  For some forums (like "Ask h2g2"), this resulted in many thousands of calculations.  What's more, it would do this before it had done the paging, doing it for every thread in the forum, as opposed to just the 25 the caller was interested in.

The "top 1…ORDER BY" varient, which is a logic equivalent, resulted in an index seek after the paging.  Sweet!

The change from "IN" to "=" is cosmetic - I just prefer "=" in this situation.