I've reviewed getrecentcomments too, as this was showing up as using far more CPU than it seemed to warrant.  There was already a Indexed View for this SP to make it go fast, but it clearly wasn't working as well as intended.
The solution to speed this up was to tweek the SP so that it filtered using the SiteId from the indexed view instead of the Forums table.  This made sure that the grouping and ordering columns both came from the view so it could use an index on those columns.  Then I changed the suplementary index on the view to index on (SiteId,DatePosted DESC) from (DatePosted DESC,UID) to support the grouping and ordering columns.

The tweak to the SP was:
	where F.SiteId = @siteid AND uc.UID like @prefix
to
	where uc.SiteId = @siteid AND uc.UID like @prefix
which is logically equivalent (the comment forum that matches the UID will never have a different SiteId to the forum associated with it)

The index change was to drop this index
CREATE INDEX IX_VUserComments_DatePosted ON VUserComments
(
	DatePosted DESC,
	uid ASC
)
and create this one:
CREATE INDEX IX_VUserComments_SiteId ON VUserComments
(
	SiteId ASC,
	DatePosted DESC
)

I ran 218 sample SPs from live and tested the output.  Both versions produced identical results.  I also checked the execution plan of inserts into the CommentForums table, the the index view update was just as efficient as the previous index.

Here are the perf improvements.

Before the changes:
Cpu:		74295
Reads:    	1372723
Duration: 	75949

After the changes
Cpu:      	16   	(99.999%  improvement)
Reads:    	18020	(98%  improvement)
Duration: 	1215  	(98%  improvement)