BBC.DNA has a post build step that calls a script called "restart memcached.bat" that stops and starts memcached.  It uses sc.exe to control the service, and a loop to wait for the service to stop before starting it again.  It will also work if you don't have a service called "memcached server" installed.
I've put the script "restart memcached.bat" in Shared so that it's easy to add to other post build steps in the future.  It probably should be run before test projects too.

The reason for this?  I got caught out today trying to find a bug that was fixed between builds because memcached had cached a version of the object between a take and build.  This will prevent that sort of problem between builds.  Also, I think it's technically sound to flush memcached between builds because your build may have changed the definition of your cached objects, which in itself will cause phantom bugs