Did the following fixes from an email from Marcus:
Striping on posts is wrong – I guess it is using the index value which is not sequential. FIXED
I hardedcoded 50 in some places – probably should be .parent. count?: - NOW NOT HARDCODED
  C:\Inetpub\wwwroot\h2g2\skins\skinsets\vanilla\boards_v2\objects\post\searchpost.xsl
I have hardcoded “forum” because I can’t find the topic name with the forum id  - all the xml is there - NOW NOT HARDCODED
When you find no terms, the error is the standard error message and looks harsh – also the search results heading is shown… - FIXED
The html title is including the whole document? - FIXED
I hardcoded color:red to the style of the searchresult guideml transform… - NOW NOT HARDCODED