Changes make BI Event Service Multithreaded.  Included all the unit test changes necessary
Also, there are changes to allow the BI event service receive NULL "risky" values from calling "predict".  If the system returns NULL, it means it doesn't know if the post is risky or not.  It will have difficulty if the post belongs to a service that has very little history, so it's unable to make a sensible decision.